Add $global:SBWebRequestTimeoutSec option
Some users are experiencing timeouts caused internally by
`Invoke-WebRequest`.  The default value for `Invoke-WebRequest`'s
`TimeoutSec` parameter is `0` (indefinite timeout), so this should
never happen.

To help gauge if there's a bug with PowerShell or not, we're adding
in this optional configuration that allows users to globally define
the timeout (in seconds) that should be provided directly to
`Invoke-WebRequest`.