Duplicate images throws an exception.
In our media retrieval logic in PackageTool, we use the -Recurse switch when trying to find a specific file.  This means there is the possibility of retrieving more than one item.  Previously, we would warn the user when this case happened and we would continue with one of the items.  For StoreBroker v2, we will throw an exception because we shouldn't be making this decision for the user.